Move poll() out into devices.c.
Make it act on a generic struct cdev included in the more specific cdev
structs.

I have used inline functions rather than macros for access to the fields
of struct cdev. It gives better type-safety and I see no reason not to
these days. (If they don't produce the same code as a macro I would like
to know why...)

Introduce a dev_t for the device numbers.