Make the IO port configurable.
For now this is in the cdev, ie. in SRAM. It's only one byte...

I am making assumptions about the layout of the different IO ports which
are not valid for all AVR chips, according to the headers. But I would
expect that where there is more than one USART on a chip, they are all
the same, so this can be handled with defines that change the layout.

With only a few exceptions, the unusual layouts correspond to the chips
using _SFR_IO instead of _SFR_MEM, which suggests to me they are older
chips in any case.