Implement generic support for fixed softc.
(softc is the proper name for device-specific configuration.)

This is more sophisticated than I entirely like, and the result is
messier than I like too. Having a macro which generates an inline
function which relies on more macros to function properly is not ideal.

We now have three possibilities for extracting a softc: static, foxed at
compile-time; inline in the device_t struct for structures which fit; or
allocated separately with a pointer in device_t.