Rename the third parameter of ioctl.
Traditionally ioctl is varargs, but I would really rather avoid that as
it passes arguments on the stack. However, there seems to be no type
which C will cast anything to without making a fuss, so the calls will
have to cast. I tried a union but you can't implicitly cast a pointer
onto a void* union member, the types have to match exactly.