Rewrite xprintf so it doesn't need a buffer.
This is a very cut-down printf, but that is appropriate here. It only
supports %d, %u, %x, %s, with %c an alias for %d. It doesn't support any
modifier characters.

This currently waits for all printing but the last section to finish
transmitting before returning. This can't be avoided without either
formatting into a buffer or using a printf task.