Remove the busy-waiting.
We don't have a timer yet, so tasks that are blocking are called
continuously and have to manage their own unblocking.

Implement a very crude timer based on an emperical measure of how long
it takes to go round an empty cycle in main(). I want to replace this
with a hardware timer really, but it works well enough for now.