Create <sys/console.h> for accessing the console.
This gives one place to compile it all out, if needed.

The console is always device 0. Replace do_write in xprintf.c with a
public function cons_write, which writes to the console and takes a char
argument.

Tidy up panic a bit (and make it work properly) and move it in with the
console functions. F_SYNC on write() is supposed to wait until the
writing has finished, but if I don't insert a delay the panic message
still doesn't get through before avr-libc shuts down the UART.