Remove ioctl.
Replace it with device-specific functions which are better able to be
inlined. The GPIO calls now compile down to optimum code; the USART
ioctls less so, but that is much less important since they're not the
main use of the USART.

Ioctl just isn't a sensible interface for this level of device; even
when it is being passed constants the code is too complex for the
compiler to optimise it away, and the packing and unpacking of boxes^W
structs takes an awful lot of code. (This is less of a concern on a real
OS where everything has to be marshalled through the syscall interface
anyway.)

The usart_* ioctls want to change to tty_* with either an addition to
the devsw or a separate ttysw for ttys. I haven't designed this yet.