Try a slower tty speed.
I would like to test slow speeds to see if adding a printf task makes
things any more responsive.

It looks like any slower than 10baud and some part of the chain gets
upset. From the tx/rx lights on the arduino I'd say it's not the AVR;
perhaps the USB chip on the board or perhaps something on the host.