Create a messages system.
The idea here is that the logging can log message numbers with
substitutions and a host-side filter can do the printf formatting. This
will not only save transmission time but also flash space in the binary.

For the moment just replace some explicit printf calls with calls to
generated MSG_* functions.

Getting lib/msg.h to build was a pig. Make doesn't identify an absolute
and a relative path for the same file, and cc -M normally generates deps
on absolute paths, so make didn't realise it could create .../lib/msg.h
using the rule for lib/msg.h. I had to change the makefile to use -MG
which creates a dep for lib/msg.h as long as that file doesn't exist.