Set up a second read buf to read ahead.
I am aiming for zero-copy. This means that, in order not to underrun on
read, the app has to give us another buffer to read into before we have
finished with the first.

This leads to a slightly weird calling sequence where we call read(2)
twice before we can process the result of the first. I'm not sure what
to do about this; perhaps it doesn't matter. However, I do think I need
some way to adjust the length of a read after it has been issued, or
perhaps this two-buffer scheme needs replacing with a low water mark.