Rework the IO system.
Devices are now opened by a particular task, and only that task can use
them. (XXX this might cause problems with console output.) That task is
the one which will be woken when IO is complete. Different tasks can own
the reading and writing halves of a device.

In order to avoid using too much SRAM for all this, restrict ourselves
to 32 tasks and store the reading/writing flags in the top bits of the
task id byte.

Change the API for read. In fact the API hasn't changed, I've just
changed the name of the function from read(2) to read_setbuf(2), as it's
very confusing having a read(2) that doesn't ever return any data. I
have also added a read_setlen(2) which updates the length of a
previously supplied buffer, so we can e.g. read the first part of a
packet to find the length and update how much we want to read based on
that. Obviously there will be problems if we've already read too much.

The generic cdev_t has now completely disappeared; cdev_t is now 'void'
like softc_t. I don't currently have any state that needs keeping for
devices like gpio that can't read or write.

poll no longer attempts to set the processor idle. I could put that
back, but I'd just as soon get rid of poll altogether. Interestingly an
entirely empty while loop doesn't work for polling; I presume the
compiler optimises it out.