Change the read API again.
Divide read(2) into two parts: read_queue and read_poll. read_queue
queues a read, and passes ownership of the buffer over to the system to
read into. real_poll checks on the status of the read, and allows the
task to wait until the read has passed a certain point before
proceeding.

This is more flexible than the previous scheme; we can achieve both
blocking reads (blocking by polling or busy-waiting atm, sleeping can be
implemented later) and O_NONBLOCK-type reads where we process whatever
we've read so far. For the latter we must pass the start of the buf to
_poll; it will return the difference between the read position and the
pointer passed.

The API is a little awkward. read_poll wants a pointer to the byte we
are interested in polling for; the caller will usually have buf+len
instead. I could make read_poll take buf and len and calculate the end,
but the effect there is that if a buf is passed that was not queued for
read then poll will return 'this buf has finished', which is odd. Maybe
I just document that passing a buf that wasn't previously passed to
_queue gives undefined results.