Begin an IIC (TWI) driver.
My attempt at writing an IRQ-driven driver straight off didn't work, so
start with a polling test program for now. It is necessary to set the
bitrate before attempting to use the bus; the datasheet doesn't spell
that out explicitly.

I have decided to use 'twi' for this specific piece of hardware and
'iic' for the i2c busses in general. This matches 'usart'/'tty'.

It is really awkward not being able to printf from interrupt context.
Possibly I could get away with disabling the interrupt we are handling
and reenabling global interrupts while I print?