Record errors from ISRs in the cdev.
cdev_rw_t gets a new member, cd_errno. This allows an ISR to pass an
error back to userland; the error will be reported next time the app
calls any function on this device.

A negative value for cd_error is an errno; a positive value is a private
error, which is reported to the console and converted into EIO.